fix: change inimation of weather containers and fix bags
