+++ weather/src/components/CalendarEvents/interface.ts
@@ -1,0 +1,4 @@
+export interface ICalendarEventsProps {
+  title: string
+  date: string
+}
