+++ weather/src/components/Weather/interface.ts
@@ -1,0 +1,6 @@
+export interface IWeatherProps {
+  temp: number
+  icon: string
+  city: string
+  weather: string
+}
