+++ weather/src/components/WeatherByDays/interface.ts
@@ -1,0 +1,6 @@
+export interface IWeatherByDaysProps {
+  date: string
+  temp: number
+  icon: string
+  wind: number
+}
