+++ weather/src/components/WeatherForcast/interface.ts
@@ -1,0 +1,7 @@
+export interface IWeatherForcastProps {
+  temp: number
+  wind: number
+  icon: string
+  description: string
+  time: string
+}
